requirements.txt
numba
numpy==1.23.0
nuscenes-devkit
opencv-python-headless
pandas
pytorch-lightning==1.6.0
scikit-image
scipy
setuptools==59.5.0
tensorboardX
